requirements.txt
# local package
-e .

# external requirements
click~=8.0.3
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

mne~=0.24.0
matplotlib~=3.4.3
setuptools~=58.0.4
numpy~=1.21.4
scipy~=1.7.1
pillow~=8.4.0
torchvision~=0.11.1